move operator classes to new submodule
