+++ src/opinf/operators/_affine.py
