+++ src/opinf/operators/_base.py
