+++ src/opinf/operators/_nonparametric.py
